requirements.txt
asgiref==3.8.1
chardet==5.2.0
colorama==0.4.6
Django==5.0.6
et-xmlfile==1.1.0
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyodbc==5.1.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
qrcode==7.4.2
reportlab==4.2.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
